refactor(wallet): export 'cip30' from index
move tests from e2e to integration

fix broken tests, add a test that goes through message passing
