+++ packages/wallet/src/index.ts
@@ -3,5 +3,5 @@ export * as KeyManagement from './KeyManagement';
 export * from './SingleAddressWallet';
 export * from './types';
 export * from './services';
-export * as util from './util';
 export * as storage from './persistence';
+export * as cip30 from './cip30';
