+++ packages/wallet/test/mocks/testKeyAgent.ts
@@ -3,9 +3,11 @@ import { KeyManagement } from '../../src';
 
 export const getPassword = jest.fn(async () => Buffer.from('password'));
 
+export const networkId = Cardano.NetworkId.testnet;
+
 export const testKeyAgent = () =>
   KeyManagement.InMemoryKeyAgent.fromBip39MnemonicWords({
     getPassword,
     mnemonicWords: KeyManagement.util.generateMnemonicWords(),
-    networkId: Cardano.NetworkId.testnet
+    networkId
   });
